requirements.txt
pandas
scikit-learn
pillow
tkinter
joblib
matplotlib
pickle